Move Lark integrations under capability package
