+++ examples/lark-kanban-control-plane-smoke.py
@@ -14,7 +14,7 @@ REPO_ROOT = Path(__file__).resolve().parents[1]
 if str(REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(REPO_ROOT))
 
-from loopx.lark_kanban import (  # noqa: E402
+from loopx.capabilities.lark.kanban import (  # noqa: E402
     CLAIM_UNCLAIMED,
     STATUS_CLAIMED,
     STATUS_DONE,
