+++ examples/lark-message-card-smoke.py
@@ -12,7 +12,7 @@ REPO_ROOT = Path(__file__).resolve().parents[1]
 if str(REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(REPO_ROOT))
 
-from loopx.lark_message_card import (  # noqa: E402
+from loopx.capabilities.lark.message_card import (  # noqa: E402
     build_lark_markdown_reply_card,
     compact_markdown,
     compact_plain_text,
