+++ loopx/capabilities/lark/__init__.py
@@ -1,0 +1,2 @@
+"""Lark/Feishu collaboration surface capability."""
+
